+++ Changelog.md
@@ -1,6 +1,7 @@
 # Change log
 
 ### vNext
+- Fix: ensure options gets up to date props [#1025](https://github.com/apollographql/react-apollo/pull/1005)
 - Fix: ensure queryRecycler exists before using it
 - MockNetworkInterface match mock requests regardless of variable order [#973](https://github.com/apollographql/react-apollo/pull/973)
 - Allow to pass removeTypenames to MockedProvider [#1001](https://github.com/apollographql/react-apollo/pull/1001)